use alternate debug / release module defns for rpycall
