+++ src/rpycall.h
@@ -1,0 +1,12 @@
+
+#ifndef RPYCALL_H
+#define RPYCALL_H
+
+namespace rpycall {
+void* rpycall_module_debug();
+void* rpycall_module_release();
+void* rpycall_methods_debug();
+void* rpycall_methods_release();
+} // namespace rpycall
+
+#endif /* RPYCALL_H */
